requirements.txt
discord.py~=2.5.0
psutil~=7.0.0
yt-dlp~=2025.1.26
sortedcontainers~=2.4.0
mysql-connector-python~=9.2.0
requests~=2.32.3